djangoProject/requirement.txt
asgiref==3.2.10
django-cors-headers==2.3.0
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
sqlparse==0.3.1
djangorestframework-jwt==1.11.0
Django==2.0
django-cors-headers==2.3.0
djangorestframework==3.9.2
cryptography==2.7
demjson



